src/pamap2_project/utils/config_loader.py
from pathlib import Path
import yaml


def load_config(config_path: str) -> dict:
    path = Path(config_path)
    if not path.exists():
        raise FileNotFoundError(f"Config non trovata: {path}")

    with path.open("r", encoding="utf-8") as f:
        cfg = yaml.safe_load(f)

    if not isinstance(cfg, dict):
        raise ValueError("Il file YAML deve contenere un dizionario alla root.")

    return cfg
