requirements.txt
numpy
pandas
pyyaml
scikit-learn
torch
torchvision
tqdm
joblib
matplotlib
